pyproject.toml
[tool.poetry]
name = "smakolyk"
version = "0.1.0"
description = ""
authors = ["\"\""]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.6"
celery = "5.2.7"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.44"
django-widget-tweaks = "^1.5.0"
pillow = "^10.4.0"
redis = "^5.0.8"
openpyxl = "^3.1.5"
black = "^24.8.0"
isort = "^5.13.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
faker = "^28.4.1"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
